Use a unified model for memory and cpu metrics
